Extract an interface for TreehouseApp
This should make it easier to write unit tests for code that
uses a TreehouseApp but doesn't need to actually download code.
