+++ redwood-treehouse-host/src/androidMain/kotlin/app/cash/redwood/treehouse/treehouseAppFactoryAndroid.kt
@@ -33,7 +33,7 @@ public fun TreehouseAppFactory(
   cacheMaxSizeInBytes: Long = 50L * 1024L * 1024L,
   concurrentDownloads: Int = 8,
   stateStore: StateStore = MemoryStateStore(),
-): TreehouseApp.Factory = TreehouseApp.Factory(
+): TreehouseApp.Factory = RealTreehouseApp.Factory(
   platform = AndroidTreehousePlatform(context),
   dispatchers = AndroidTreehouseDispatchers(),
   httpClient = httpClient.asZiplineHttpClient(),
